added the <asset> class
the asset class is used for doing operations on pieces
of capital which do not have the same unit. adding capital
or another asset to an asset will attempt to consolidate the
capital as much as possible inside of the asset. all algebraic
operations which are possible on bits of capital are also
applicable on assets